Refactor(cookie-log): answer record fetch 시점 수정
